--- __init__.py
@@ -1,4 +1,0 @@
-from .app import create_app
-
-APP = create_app()
-
