+++ appy/__init__.py
@@ -1,0 +1,4 @@
+from .app import flask_app
+
+APP = flask_app()
+
